Move independent utilities into lib/application_utils.rb
