+++ application.rb
@@ -25,6 +25,8 @@ configure do
 end
 
 helpers do
+  include ApplicationUtils
+
   def versioned_stylesheet(style)
     "/#{style}.css?" + File.mtime(File.join(settings.public_folder, "scss", "#{style}.scss")).to_i.to_s
   end
@@ -69,30 +71,6 @@ helpers do
     end
   end
 
-  def get_months
-    [
-      {:index => 1, :name => "Jan"},
-      {:index => 2, :name => "Feb"},
-      {:index => 3, :name => "Mar"},
-      {:index => 4, :name => "Apr"},
-      {:index => 5, :name => "May"},
-      {:index => 6, :name => "Jun"},
-      {:index => 7, :name => "Jul"},
-      {:index => 8, :name => "Aug"},
-      {:index => 9, :name => "Sep"},
-      {:index => 10, :name => "Oct"},
-      {:index => 11, :name => "Nov"},
-      {:index => 12, :name => "Dec"}
-    ]
-  end
-
-  def get_days
-    (1..31)
-  end
-
-  def get_years
-    (1900..2048)
-  end
 end
 
 before do
